GETP-216 refactor: domain 모듈 패키지 구조가 바운디드 컨텍스트를 나타내도록 변경
